split up api tests
